requirements.txt
confluent-kafka
requests
clickhouse-connect
psycopg2-binary
web3
mypy
flake8
prometheus_client